requirements.txt
requests>=2.31.0
uvicorn>=0.30.0 # if running local FastAPI-based servers in dev
fastapi>=0.115.0 # optional: aligns with typical Marker API stacks
pypdf>=4.2.0
tiktoken>=0.5.0

marker-pdf>=1.10.1
streamlit>=1.29.0
openai>=1.3.0
python-dotenv>=1.0.0
pytest>=8.3.0
